doc type exported
